refactor(docs): drop docs-kit namespace for lib/docs utilities
