--- apps/docs/src/docs-kit/chrome-events.ts
@@ -1,5 +1,0 @@
-/** Window event names used by the docs sidebar burger integration. */
-export const docsKitBurgerEvents = {
-	TOGGLE_MENU: 'toggle-menu',
-	CLOSE_MENU: 'close-menu',
-} as const;
