--- apps/docs/src/docs-kit/docs-mdx.types.ts
@@ -1,4 +1,0 @@
-import type { EcoComponent } from '@ecopages/core';
-
-/** Imported MDX default export from the ecopages-jsx loader. */
-export type DocsMdxComponent = EcoComponent<Record<string, unknown>>;
