+++ apps/docs/src/layouts/docs-layout/docs-layout.tsx
@@ -6,8 +6,8 @@ import { Banner } from '@/components/banner/banner';
 import { CodeTabs } from '@/components/code-tabs';
 import { docsNav, serializeDocsPaginationData } from '@/lib/content-nav';
 import { BaseLayout } from '@/layouts/base-layout';
-import { resolveDocsBreadcrumb } from '@/docs-kit/resolve-docs-breadcrumb';
-import { getDocsLlmUrl } from '@/docs-kit/docs-llm-url';
+import { resolveDocsBreadcrumb } from '@/lib/docs/resolve-docs-breadcrumb';
+import { getDocsLlmUrl } from '@/lib/docs/docs-llm-url';
 import { DocsBar } from './docs-bar';
 import { DocsPagination } from './components/docs-pagination';
 import { DocsSidebar } from './components/navigation';
