+++ apps/docs/src/lib/docs/docs-llm-url.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest';
-import { getDocsLlmUrlFromPathname } from '@/docs-kit/docs-llm-url';
+import { getDocsLlmUrlFromPathname } from '@/lib/docs/docs-llm-url';
 
 test('getDocsLlmUrlFromPathname maps docs paths to markdown URLs', () => {
 	expect(getDocsLlmUrlFromPathname('/docs/getting-started/introduction')).toBe(
