+++ apps/docs/vitest.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
 	},
 	test: {
 		environment: 'node',
-		include: ['scripts/**/*.test.ts', 'src/docs-kit/**/*.test.ts'],
+		include: ['scripts/**/*.test.ts', 'src/lib/docs/**/*.test.ts'],
 		setupFiles: ['./vitest.setup.ts'],
 	},
 });
