+++ examples/docs-starter/src/components/copy-for-llm/copy-for-llm.script.tsx
