+++ examples/docs-starter/src/components/copy-for-llm/index.tsx
