+++ examples/docs-starter/src/content/docs/getting-started/next-steps.mdx
@@ -8,4 +8,4 @@ order: 2
 
 - Copy this example into your own app.
 - Add MDX pages under `src/content/docs/<section>/`.
-- Add shared JSX in `src/docs-kit/mdx-components.ts` and pass it through `getDocsMdxComponents()`.
+- Add shared JSX in `src/lib/docs/mdx-components.ts` and pass it through `getDocsMdxComponents()`.
