--- examples/docs-starter/src/docs-kit/mdx/docs-mdx.types.ts
@@ -1,3 +1,0 @@
-import type { EcoComponent } from '@ecopages/core';
-
-export type DocsMdxComponent = EcoComponent<Record<string, unknown>>;
