+++ examples/docs-starter/src/lib/docs/resolve-from-catch-all.ts
