+++ vitest.browser.config.ts
@@ -1,7 +1,16 @@
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 import { defineProject } from 'vitest/config';
 import { playwright } from '@vitest/browser-playwright';
 
+const rootDir = path.dirname(fileURLToPath(import.meta.url));
+
 export default defineProject({
+	resolve: {
+		alias: {
+			'@': path.resolve(rootDir, 'apps/docs/src'),
+		},
+	},
 	test: {
 		name: 'browser',
 		include: [
